KRACOEUS-6955 move *Acl* classes to kra protocol auth for later repackaging
