+++ coeus-code/src/main/java/org/kuali/kra/protocol/auth/UnitAclLoadService.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.kuali.kra.service;
+package org.kuali.kra.protocol.auth;
 
 import org.kuali.coeus.sys.framework.auth.perm.Permissionable;
