+++ coeus-code/src/main/java/org/kuali/kra/protocol/auth/UnitAclLoadServiceImpl.java
@@ -13,13 +13,13 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.kuali.kra.service.impl;
+package org.kuali.kra.protocol.auth;
 
 import org.apache.commons.lang3.StringUtils;
 import org.kuali.coeus.sys.framework.auth.SystemAuthorizationService;
 import org.kuali.coeus.sys.framework.auth.perm.KcAuthorizationService;
 import org.kuali.coeus.sys.framework.auth.perm.Permissionable;
-import org.kuali.kra.service.UnitAclLoadService;
+import org.kuali.kra.protocol.auth.UnitAclLoadService;
 import org.kuali.rice.kim.api.role.Role;
 import org.kuali.rice.kim.api.role.RoleMembership;
 import org.kuali.rice.kim.api.role.RoleService;
