+++ coeus-code/src/main/java/org/kuali/kra/protocol/onlinereview/ProtocolOnlineReviewBase.java
@@ -22,7 +22,7 @@ import org.kuali.coeus.common.framework.rolodex.Rolodex;
 import org.kuali.coeus.common.permissions.impl.PermissionableKeys;
 import org.kuali.coeus.sys.framework.auth.perm.Permissionable;
 import org.kuali.coeus.sys.framework.model.KcPersistableBusinessObjectBase;
-import org.kuali.kra.UnitAclLoadable;
+import org.kuali.kra.protocol.auth.UnitAclLoadable;
 import org.kuali.kra.infrastructure.Constants;
 import org.kuali.kra.infrastructure.RoleConstants;
 import org.kuali.kra.protocol.ProtocolBase;
