+++ coeus-code/src/main/resources/org/kuali/kra/CoreSpringBeans.xml
@@ -1681,7 +1681,7 @@
         </property>
     </bean>
 
-    <bean id="unitAclLoadService" class="org.kuali.kra.service.impl.UnitAclLoadServiceImpl">
+    <bean id="unitAclLoadService" class="org.kuali.kra.protocol.auth.UnitAclLoadServiceImpl">
         <property name="kraAuthorizationService" ref="kcAuthorizationService" />
         <property name="roleManagementService" ref="roleService" />
         <property name="systemAuthorizationService" ref="systemAuthorizationService" />
